consolidate shared naming functionality
