+++ packages/naming.go/mappings.ts
