+++ packages/naming.go/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@azure-tools/naming.go",
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "build": "tsc -p ."
+  },
+  "devDependencies": {
+    "typescript": "~5.1.3"
+  }
+}
