+++ packages/typespec-go/src/tcgcadapter/clients.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { capitalize, ensureNameCase, getEscapedReservedName, uncapitalize } from './namer.js';
+import { capitalize, ensureNameCase, getEscapedReservedName, uncapitalize } from '../../../naming.go/naming.js';
 import { isTypePassedByValue, typeAdapter } from './types.js';
 import * as go from '../../../codemodel.go/gocodemodel.js';
 import * as tcgc from '@azure-tools/typespec-client-generator-core';
