--- packages/typespec-go/src/tcgcadapter/mappings.ts
@@ -1,134 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export const CommonAcronyms = [
-  '^aad$',
-  '^acl', // to match Acls
-  'aes\\d+',
-  '^api$',
-  '^ascii$',
-  '^cpu$',
-  '^crc\\d*$',
-  '^css$',
-  '^csv$',
-  '^dns$',
-  '^eof$',
-  'gcm',
-  '^guid$',
-  '^html$',
-  '^http[s|\\d]?$',
-  '^jpeg$',
-  '^(id)s?$',
-  '^(ipc)onfig',
-  '^(ip)s?$',
-  '^(ips)ec',
-  '^(ip)[vV]\\d',
-  '^json$',
-  '^lhs$',
-  '^mbcs$',
-  '^md5$',
-  '^(od)ata$',
-  '^pdf$',
-  '^png$',
-  '^qps$',
-  '^ram$',
-  '^rfc\\d*$',
-  '^rhs$',
-  '^rpc$',
-  '^sha\\d*$',
-  '^(sku)[sS]?$',
-  '^sla$',
-  '^smtp$',
-  '^sql$',
-  '^ssh$',
-  '^ssl$',
-  '^tcp$',
-  '^tiff$',
-  '^tls$',
-  '^ttl$',
-  '^udp$',
-  '^ui$',
-  '^uid$',
-  '^uuid$',
-  '^uri$',
-  '^url$',
-  '^utc$',
-  '^utf\\d*$',
-  '^(vm)s?$',
-  '^(vpn)s?$',
-  '^xml$',
-  '^xsrf$',
-  '^xss$'
-];
-
-export const ReservedWords = [
-  // Reserved keywords -- list retrieved from http://golang.org/ref/spec#Keywords
-  'break', 'default', 'func', 'interface', 'select',
-  'case', 'defer', 'go', 'map', 'struct',
-  'chan', 'else', 'goto', 'package', 'switch',
-  'const', 'fallthrough', 'if', 'range', 'type',
-  'continue', 'for', 'import', 'return', 'var',
-
-  // Reserved predeclared identifiers -- list retrieved from http://golang.org/ref/spec#Predeclared_identifiers
-  'bool', 'byte',
-  'complex64', 'complex128',
-  'error',
-  'float32', 'float64',
-  'int', 'int8', 'int16', 'int32', 'int64',
-  'rune', 'string',
-  'uint', 'uint8', 'uint16', 'uint32', 'uint64',
-  'uintptr',
-
-  'true', 'false', 'iota',
-
-  'nil',
-
-  'append', 'cap', 'close', 'complex', 'copy', 'delete', 'imag', 'len', 'make', 'new', 'panic', 'print', 'println', 'real', 'recover',
-
-
-  // Reserved packages -- list retrieved from http://golang.org/pkg/
-  // -- Since package names serve as partial identifiers, exclude the standard library
-  'archive', 'tar', 'zip',
-  'bufio',
-  'builtin',
-  'bytes',
-  'compress', 'bzip2', 'flate', 'gzip', 'lzw', 'zlib',
-  'container', 'heap', 'list', 'ring',
-  'context',
-  'crypto', 'aes', 'cipher', 'des', 'dsa', 'ecdsa', 'ed25519', 'elliptic', 'hmac', 'md5', 'rand', 'rc4', 'rsa', 'sha1', 'sha256', 'sha512', 'subtle', 'tls', 'x509', 'pkix',
-  'database', 'sql', 'driver',
-  'debug', 'dwarf', 'elf', 'gosym', 'macho', 'pe', 'plan9obj',
-  'encoding', 'ascii85', 'asn1', 'base32', 'base64', 'binary', 'csv', 'gob', 'hex', 'json', 'pem', 'xml',
-  'errors',
-  'expvar',
-  'flag',
-  'fmt',
-  'go', 'ast', 'build', 'constant', 'doc', 'format', 'importer', 'parser', 'printer', 'scanner', 'token', 'types',
-  'hash', 'adler32', 'crc32', 'crc64', 'fnv',
-  'html', 'template',
-  'image', 'color', 'palette', 'draw', 'gif', 'jpeg', 'png',
-  'index', 'suffixarray',
-  'io', 'ioutil',
-  'log', 'syslog',
-  'math', 'big', 'bits', 'cmplx', 'rand',
-  'mime', 'multipart', 'quotedprintable',
-  'net', 'http', 'cgi', 'cookiejar', 'fcgi', 'httptest', 'httptrace', 'httputil', 'pprof', 'mail', 'rpc', 'jsonrpc', 'smtp', 'textproto', 'url',
-  'os', 'exec', 'signal', 'user',
-  'path', 'filepath',
-  'plugin',
-  'reflect',
-  'regexp', 'syntax',
-  'runtime', 'cgo', 'debug', 'msan', 'pprof', 'race', 'trace',
-  'sort',
-  'strconv',
-  'strings',
-  'sync', 'atomic',
-  'syscall', 'js',
-  'testing', 'iotest', 'quick',
-  'text', 'scanner', 'tabwriter', 'template', 'parse',
-  'time',
-  'unicode', 'utf16', 'utf8',
-  'unsafe'
-];
