+++ packages/typespec-go/src/tcgcadapter/types.ts
@@ -3,7 +3,7 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { capitalize } from './namer.js';
+import { capitalize } from '../../../naming.go/naming.js';
 import * as go from '../../../codemodel.go/gocodemodel.js';
 import * as tcgc from '@azure-tools/typespec-client-generator-core';
 import * as tsp from '@typespec/compiler';
