+++ rush.json
@@ -121,6 +121,12 @@
       "shouldPublish": true
     },
     {
+      "packageName": "@azure-tools/naming.go",
+      "projectFolder": "packages/naming.go",
+      "reviewCategory": "production",
+      "shouldPublish": true
+    },
+    {
       "packageName": "@autorest/go",
       "projectFolder": "packages/autorest.go",
       "reviewCategory": "production",
